Remove code in __init__.py files
Enables manubot CLI to be called without running these files and
triggering the imports
