+++ tests/test_manubot.py
@@ -3,7 +3,7 @@ import subprocess
 
 import pytest
 
-from manubot.process import (
+from manubot.process.process import (
     add_author_affiliations,
     read_jsons,
 )
